fix: nested json table string length
In cases where we are parsing nested json with lengthy key strings the
combination of key strings in the breadcrumb based table name string
would result in an error when trying to train the model due to a maximum
json filename length of 128 chars. Here we utilize the already existing
make_suffix function (which ensures no sanitized_str collisions) and we
truncate the length of the table name to ensure we do not exceed the 128
char limit.