feat: store json table mapping in debug_summary
When nested json structures are extracted to individual dataframes they
will be written to resulting individual tables. Since we are creating
table names with the convention <producer_table>_invented_<uuid> we now
store the mapping details for tables in the producer table's debug
summary, as well as the json breadcrumb path in each invented table's
individual debug summary.