fix: nested json table string length (#135)
* fix: nested json table string length

In cases where we are parsing nested json with lengthy key strings the
combination of key strings in the breadcrumb based table name string
would result in an error when trying to train the model due to a maximum
json filename length of 128 chars. Here we utilize the already existing
make_suffix function (which ensures no sanitized_str collisions) and we
truncate the length of the table name to ensure we do not exceed the 128
char limit.

* feat: remove json breadcrumbs from table names

Instead of sanitizing the full json path to an invented table, we create
a table name based on the producer table with a uuid to prevent
collisions, and we map the full json path to the unique table string.

* feat: store json table mapping in debug_summary

When nested json structures are extracted to individual dataframes they
will be written to resulting individual tables. Since we are creating
table names with the convention <producer_table>_invented_<uuid> we now
store the mapping details for tables in the producer table's debug
summary, as well as the json breadcrumb path in each invented table's
individual debug summary.